Disable availability nag modal on members dashboard
Extracts AvailabilityNagModal to its own component file and removes it
from the dashboard render. The banner reminder remains. Re-enable by
importing and adding <AvailabilityNagModal /> back to the dashboard.
